requirements.txt
paho-mqtt==1.6.1
python-dotenv==0.19.2
smbus2==0.4.1
termcolor==1.1.0
